extracted column out of board into seperate component
